refatorando um pouco o código
