+++ ds2m/caio_palermo_lemos/css/chats/chatsContainerStyle.css
@@ -1,0 +1,58 @@
+.chat:hover {
+  cursor: pointer;
+}
+
+.chat__image {
+  width: 64px;
+  height: 64px;
+  border-radius: 100%;
+}
+
+.chats__user-interactions i,
+span:hover {
+  cursor: pointer;
+}
+
+.chats__chats-container {
+  display: flex;
+  flex-direction: column;
+  max-height: 80vh;
+  overflow-y: auto;
+}
+
+.chat {
+  display: grid;
+  grid-template-areas: "image content";
+  grid-template-columns: auto 1fr;
+  align-items: center;
+  margin-right: 12px;
+}
+
+.chat:hover {
+  background-color: var(--secondary-color);
+  border: none;
+}
+
+.chat__image {
+  margin-left: 16px;
+  width: 68px;
+  height: 68px;
+}
+
+.chat__content {
+  padding: 16px 0px 16px 0px;
+  margin-left: 28px;
+  border-top: 1px solid #384045;
+  border-bottom: 1px solid #384045;
+}
+
+.chat__name {
+  color: var(--text-color);
+  font-size: 2rem;
+  margin-bottom: 5px;
+}
+
+.chat__text {
+  color: var(--text-subcolor);
+  font-size: 1.25rem;
+}
