+++ ds2m/caio_palermo_lemos/css/chats/chatsProfileOptionsStyle.css
@@ -1,0 +1,45 @@
+.chats__profile-options {
+  height: 10vh;
+  display: grid;
+  grid-template-columns: 1fr auto auto auto auto;
+  gap: 48px;
+  padding: 12px 24px 12px 24px;
+  align-items: center;
+  background-color: var(--secondary-color);
+}
+
+.chats__profile-options-mobile {
+  display: none;
+}
+
+#profile-options__profile {
+  background-color: var(--text-color);
+  border-radius: 100%;
+  width: 64px;
+  height: 64px;
+  text-align: center;
+  line-height: 64px;
+  color: var(--secondary-color);
+}
+
+.chats__profile-options > i,
+img {
+  font-size: 2rem;
+  width: 32px;
+  height: 32px;
+  color: var(--text-color);
+}
+
+.profile-options-mobile__text {
+  font-size: 1.5rem;
+  color: var(--text-color);
+  font-weight: bold;
+}
+
+.profile-options-mobile__icons > i,
+img {
+  font-size: 1.5rem;
+  width: 32px;
+  height: 32px;
+  color: var(--text-color);
+}
