+++ ds2m/caio_palermo_lemos/css/chats/chatsStyle.css
@@ -1,0 +1,4 @@
+.chats {
+  background-color: var(--primary-color);
+  border-right: 1px solid #384045;
+}
