+++ ds2m/caio_palermo_lemos/css/displays.css
@@ -1,0 +1,11 @@
+.none {
+  display: none !important;
+}
+
+.block {
+  display: block !important;
+}
+
+.flex {
+  display: flex !important;
+}
