+++ ds2m/caio_palermo_lemos/css/messages/messagesBackgroundStyle.css
@@ -1,0 +1,8 @@
+.messages__background-default {
+  width: 100%;
+  height: 99.9%;
+}
+
+.messages__background-default {
+  display: block;
+}
