+++ ds2m/caio_palermo_lemos/css/messages/messagesStyle.css
@@ -1,0 +1,3 @@
+.messages {
+  background-color: var(--primary-color);
+}
