+++ ds2m/caio_palermo_lemos/css/return.css
@@ -1,0 +1,8 @@
+.return {
+  display: none;
+  position: absolute;
+  left: 20px;
+  top: 27px;
+  font-size: 2rem;
+  color: var(--text-color);
+}
