+++ ds2m/caio_palermo_lemos/css/style.css
@@ -1,3 +1,7 @@
+html {
+  scroll-behavior: smooth;
+}
+
 body {
   font-family: "Open Sans", sans-serif;
 }
@@ -11,365 +15,4 @@ main {
   display: grid;
   grid-template-columns: 1fr 1.5fr;
   min-height: 100vh;
-}
-
-.chats {
-  background-color: var(--primary-color);
-  border-right: 1px solid #384045;
-}
-
-.chat:hover {
-  cursor: pointer;
-}
-
-.chats__profile-options {
-  height: 10vh;
-  display: grid;
-  grid-template-columns: 1fr auto auto auto auto;
-  gap: 48px;
-  padding: 12px 24px 12px 24px;
-  align-items: center;
-  background-color: var(--secondary-color);
-}
-
-.chats__profile-options-mobile {
-  display: none;
-}
-
-#profile-options__profile {
-  background-color: var(--text-color);
-  border-radius: 100%;
-  width: 64px;
-  height: 64px;
-  text-align: center;
-  line-height: 64px;
-  color: var(--secondary-color);
-}
-
-.chats__profile-options > i,
-img {
-  font-size: 2rem;
-  width: 32px;
-  height: 32px;
-  color: var(--text-color);
-}
-
-.profile-options-mobile__text {
-  font-size: 1.5rem;
-  color: var(--text-color);
-  font-weight: bold;
-}
-
-.profile-options-mobile__icons > i,
-img {
-  font-size: 1.5rem;
-  width: 32px;
-  height: 32px;
-  color: var(--text-color);
-}
-
-.chats__user-interactions {
-  display: none;
-}
-
-.chat__image {
-  width: 64px;
-  height: 64px;
-  border-radius: 100%;
-}
-
-.chats__filter {
-  display: flex;
-  gap: 32px;
-  justify-content: center;
-  align-items: center;
-  margin: 12px;
-  margin-bottom: 24px;
-}
-
-.filter__input-container {
-  position: relative;
-}
-
-.filter__input {
-  width: 35vw;
-  height: 50px;
-  background-color: var(--secondary-color);
-  border-radius: 8px;
-  border: none;
-  color: var(--text-color);
-  font-size: 1.3rem;
-  padding-left: 100px;
-}
-
-.filter__input::placeholder {
-  color: var(--text-subcolor);
-}
-
-#filter__icon-search {
-  position: absolute;
-  top: 15px;
-  left: 25px;
-  color: var(--text-color);
-  font-size: 1.25rem;
-}
-
-.messages {
-  background-color: var(--primary-color);
-}
-
-.messages__background-default {
-  width: 100%;
-  height: 99.9%;
-}
-
-.messages__background-default {
-  display: block;
-}
-
-.messages__receiver {
-  display: none;
-}
-
-.messages__conversations-container {
-  display: none;
-}
-
-.messages__interactions {
-  display: none;
-}
-
-.chats__user-interactions i,
-span:hover {
-  cursor: pointer;
-}
-
-.user-interactions__conversations::after {
-  content: "";
-  display: block;
-  border: 1px solid var(--text-color);
-}
-
-.chats__chats-container {
-  display: flex;
-  flex-direction: column;
-}
-
-.chat {
-  display: grid;
-  grid-template-areas: "image content";
-  grid-template-columns: auto 1fr;
-  align-items: center;
-  margin-right: 12px;
-}
-
-.chat:hover {
-  background-color: var(--secondary-color);
-  border: none;
-}
-
-.chat__image {
-  margin-left: 16px;
-  width: 68px;
-  height: 68px;
-}
-
-.chat__content {
-  padding: 16px 0px 16px 0px;
-  margin-left: 28px;
-  border-top: 1px solid #384045;
-  border-bottom: 1px solid #384045;
-}
-
-/* .chat__content:hover {
-  border: transparent;
-} */
-
-.chat__name {
-  color: var(--text-color);
-  font-size: 2rem;
-  margin-bottom: 5px;
-}
-
-.chat__text {
-  color: var(--text-subcolor);
-  font-size: 1.25rem;
-}
-
-.messages__receiver {
-  height: 10vh;
-  display: flex;
-  justify-content: space-between;
-  align-self: center;
-  background-color: var(--secondary-color);
-}
-
-.receiver__profile {
-  display: grid;
-  grid-template-columns: auto 1fr auto;
-  align-items: center;
-  width: 100%;
-}
-
-.profile__image {
-  width: 64px;
-  height: 64px;
-  border-radius: 100%;
-  margin: 0px 24px;
-}
-
-.profile__name {
-  font-size: 2rem;
-  color: var(--text-color);
-}
-
-.receiver__options {
-  display: flex;
-  align-items: center;
-  gap: 64px;
-  margin: 0px 48px;
-}
-
-.receiver__options i {
-  font-size: 1.75rem;
-  color: var(--text-color);
-}
-
-.messages__conversations-container {
-  padding-top: 25px;
-  min-height: 80vh;
-  max-height: 80vh;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-  overflow: auto;
-}
-
-.messages__conversations {
-  padding-top: 25px;
-  display: flex;
-  flex-direction: column;
-  gap: 18px;
-}
-
-.message-sender-container {
-  display: flex;
-  justify-content: flex-start;
-}
-
-.message__sender {
-  margin-left: 100px;
-  padding: 12px;
-  display: grid;
-  justify-content: start;
-  grid-template-columns: auto auto;
-  position: relative;
-  max-width: 60%;
-  width: fit-content;
-  height: auto;
-  background-color: var(--secondary-color);
-  border-radius: 8px;
-  color: var(--text-color);
-}
-
-.tri-right.left-top:after {
-  content: " ";
-  position: absolute;
-  width: 0;
-  height: 0;
-  left: -10px;
-  right: auto;
-  top: 0px;
-  bottom: auto;
-  border: 12px solid;
-  border-color: var(--secondary-color) transparent transparent transparent;
-}
-
-.message-receiver-container {
-  display: flex;
-  justify-content: flex-end;
-}
-
-.message__receiver {
-  margin-right: 100px;
-  padding: 12px;
-  display: flex;
-  position: relative;
-  max-width: 60%;
-  width: fit-content;
-  height: auto;
-  background-color: #005c4b;
-  border-radius: 8px;
-  color: var(--text-color);
-}
-
-.tri-right.right-top:after {
-  content: " ";
-  position: absolute;
-  width: 0;
-  height: 0;
-  left: auto;
-  right: -10px;
-  top: 0px;
-  bottom: auto;
-  border: 12px solid;
-  border-color: #005c4b transparent transparent transparent;
-}
-
-.message__content {
-  padding: 8px;
-  display: flex;
-  align-items: flex-start;
-  font-size: 1.5rem;
-}
-
-.message__time {
-  display: flex;
-  align-items: flex-end;
-  margin-left: 1px;
-  margin-bottom: 2px;
-  margin-right: 4px;
-  font-size: 1rem;
-}
-
-.messages__interactions {
-  height: 10vh;
-  background-color: var(--secondary-color);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 24px;
-}
-
-.messages__interactions i {
-  font-size: 2rem;
-  color: var(--text-subcolor);
-}
-
-.interactions__send-mesage {
-  width: 80%;
-  height: 70%;
-  border: none;
-  border-radius: 20px;
-  background-color: #293942;
-  color: var(--text-color);
-  font-size: 1.5rem;
-  padding-left: 30px;
-}
-
-.interactions__send-mesage::placeholder {
-  color: var(--text-subcolor);
-}
-
-.none {
-  display: none !important;
-}
-
-.block {
-  display: block !important;
-}
-
-.flex {
-  display: flex !important;
 }
