+++ ds2m/caio_palermo_lemos/index.html
@@ -16,10 +16,24 @@
     />
     <link rel="icon" type="imagem/png" href="./img/whatsapp.png" />
     <script src="./js/app.js" defer type="module"></script>
-    <script src="./js/adicionarAtivo.js" defer></script>
     <link rel="stylesheet" href="./css/reset.css" />
     <link rel="stylesheet" href="./css/variables.css" />
+    <link rel="stylesheet" href="./css/displays.css" />
     <link rel="stylesheet" href="./css/style.css" />
+    <link rel="stylesheet" href="./css/chats/chatsStyle.css" />
+    <link rel="stylesheet" href="./css/chats/chatsProfileOptionsStyle.css" />
+    <link rel="stylesheet" href="./css/chats/chatsUserInteractionsStyle.css" />
+    <link rel="stylesheet" href="./css/chats/chatsFilterStyle.css" />
+    <link rel="stylesheet" href="./css/chats/chatsContainerStyle.css" />
+    <link rel="stylesheet" href="./css/return.css" />
+    <link rel="stylesheet" href="./css/messages/messagesStyle.css" />
+    <link rel="stylesheet" href="./css/messages/messagesBackgroundStyle.css" />
+    <link rel="stylesheet" href="./css/messages/messagesReceiverStyle.css" />
+    <link
+      rel="stylesheet"
+      href="./css/messages/messagesConversationsStyle.css"
+    />
+    <link rel="stylesheet" href="./css/messages/messagesInteractions.css" />
     <link rel="stylesheet" href="./css/responsive.css" />
     <title>WhatsApp</title>
   </head>
@@ -76,31 +90,10 @@
           alt=""
           class="messages__background-default"
         />
-        <div class="messages__receiver">
-          <!-- <div class="receiver__profile">
-            <img src="./img/branca.jpg" alt="" class="profile__image" />
-            <p class="profile__name">Branca</p>
-          </div>
-          <div class="receiver__options">
-            <i class="fa-solid fa-magnifying-glass" id="options__search"></i>
-            <i class="fas fa-ellipsis-v" id="options__more-options"></i>
-          </div> -->
-        </div>
+        <div class="messages__receiver"></div>
+        <i class="fas fa-laugh return"></i>
         <div class="messages__conversations-container">
-          <div class="messages__conversations">
-            <!-- <div class="message-sender-container">
-              <div class="message__sender tri-right left-top">
-                <p class="message__content">Chuchu, suas telas estão prontas</p>
-                <span class="message__time">11:21 AM</span>
-              </div>
-            </div>
-            <div class="message-receiver-container">
-              <div class="message__receiver tri-right right-top">
-                <p class="message__content">Que belezinha!</p>
-                <span class="message__time">13:54 PM</span>
-              </div>
-            </div> -->
-          </div>
+          <div class="messages__conversations"></div>
         </div>
         <div class="messages__interactions">
           <i class="fas fa-laugh"></i>
